Minor fix (#43)
* Use correct type for device ID.

* Remove unused RMT_TODO.

+++ agent/rmt_agent.h
@@ -7,6 +7,8 @@ extern "C" {
 #endif
 /* *INDENT-ON* */
 
+#include <stdint.h>
+
 /**
  * @file
  * RMT agent library API.
@@ -32,7 +34,7 @@ typedef struct fileinfo_func {
 
 typedef struct _rmt_agent_cfg {
     char *net_interface;             // NULL for default interface chosen by RMT agent
-    int device_id;                   // 0 for default device ID generated from MAC
+    uint64_t device_id;              // 0 for default device ID generated from MAC
     int domain_id;                   // 0 for default domain ID 0
     unsigned long datainfo_val_size; // 0 for default size 256
     unsigned long devinfo_size;      // 0 for default size 1024
